[SRM-10274]: CVNG V2 plan creator upgrade (#33974)
