--- 300-cv-nextgen/src/main/java/io/harness/cvng/cdng/services/impl/CVNGPlanCreator.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2021 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
- */
-
-package io.harness.cvng.cdng.services.impl;
-
-import io.harness.advisers.manualIntervention.ManualInterventionAdviserRollbackParameters;
-import io.harness.advisers.manualIntervention.ManualInterventionAdviserWithRollback;
-import io.harness.annotations.dev.HarnessTeam;
-import io.harness.annotations.dev.OwnedBy;
-import io.harness.cvng.cdng.beans.CVNGStepType;
-import io.harness.plancreator.steps.GenericStepPMSPlanCreator;
-import io.harness.pms.contracts.advisers.AdviserObtainment;
-import io.harness.pms.contracts.execution.failure.FailureType;
-import io.harness.pms.yaml.YamlField;
-import io.harness.utils.TimeoutUtils;
-import io.harness.yaml.core.failurestrategy.FailureStrategyActionConfig;
-import io.harness.yaml.core.failurestrategy.manualintervention.ManualInterventionFailureActionConfig;
-
-import com.google.common.collect.Sets;
-import com.google.protobuf.ByteString;
-import java.util.Set;
-@OwnedBy(HarnessTeam.CV)
-public class CVNGPlanCreator extends GenericStepPMSPlanCreator {
-  public static final Set<String> CVNG_SUPPORTED_TYPES = Sets.newHashSet(CVNGStepType.CVNG_VERIFY.getDisplayName());
-  @Override
-  public Set<String> getSupportedStepTypes() {
-    return CVNG_SUPPORTED_TYPES;
-  }
-
-  @Override
-  protected AdviserObtainment getManualInterventionAdviserObtainment(Set<FailureType> failureTypes,
-      AdviserObtainment.Builder adviserObtainmentBuilder, ManualInterventionFailureActionConfig actionConfig,
-      FailureStrategyActionConfig actionUnderManualIntervention, YamlField currentField) {
-    return adviserObtainmentBuilder.setType(ManualInterventionAdviserWithRollback.ADVISER_TYPE)
-        .setParameters(ByteString.copyFrom(kryoSerializer.asBytes(
-            ManualInterventionAdviserRollbackParameters.builder()
-                .applicableFailureTypes(failureTypes)
-                .timeoutAction(toRepairAction(actionUnderManualIntervention))
-                .timeout((int) TimeoutUtils.getTimeoutInSeconds(actionConfig.getSpecConfig().getTimeout(), 0))
-                .build())))
-        .build();
-  }
-}
